Applying the patch of Yanqin Jin <y7jin@eng.ucsd.edu> in btree_impl.cpp
In sm/btree_impl.cpp:4044 (_lookup function), a btrec_t rec is defined and used later to capture the desired key-value pair on the btree page. Then in sm/btree_impl.cpp:4070, the lock request may fail due to the fact that it is a conditional lock request. If it does fail, we will unfix child (leaf) and parent altogether. Then we unconditionally ask for the same lock again. In the meanwhile the buffer pool frame which used to host that particular btree page may well have been paged out by another thread, especially when the buffer pool size is small. Consequently, when the second lock request succeeds and we reach sm/btree_impl.cpp:4100, although child now again fixes that desired btree page, the variable rec still points to the slot on the old buffer frame. This will cause data corruption when we enter the branch in sm/btree_impl.cpp:4146. I have fixed this bug by setting the rec variable again immediately after fixing child and it seems to work for me. We also have to update the cursor. This problem will occur in situations when you have really small buffer pool size.